3_function_move_data_to_dataset_big_query_with_sql_query/requirements.txt
google-cloud-bigquery
google-cloud-kms
pandas-gbq